requirements.txt
Flask==3.1.1
transformers==4.53.0
torch==2.7.1
torchvision==0.22.1
Pillow==11.2.1
faiss-cpu==1.11.0
jinja2==3.1.6
requests==2.32.4
